box: initialisation
This commit does mulitple things:

1.
We have Config structure that reads json config from a preset location.
It holds these parameters:

* Box ID
* Drone ID
* Initialisation Flag
* Landing location of the box

The config can be read and written to the predefined location.

2.
Hardcoded topics are not written in the code. Instead they are placed in
the utils/topics.py file.

3.
The box has 2 stages:
* Setup
    - There is only 1 publisher and 1 subscriber. The publisher
publishes the box's id every second and waits for a response on the
subscriber.

* Working
    - The publisher, subscriber and timer from the setup stage are
destroied and new subscribers and publishers are created.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>